+++ R/Loggers.R
@@ -84,7 +84,7 @@ addDefaultConsoleLogger <- function() {
 #'
 #' @export
 addDefaultFileLogger <- function(fileName) {
-  registerLogger(createLogger(name = "DEFAULT",
+  registerLogger(createLogger(name = "DEFAULT_FILE_LOGGER",
                               threshold = "TRACE",
                               appenders = list(createFileAppender(layout = layoutParallel,
                                                                   fileName = fileName))))
@@ -123,10 +123,30 @@ addDefaultFileLogger <- function(fileName) {
 #'
 #' @export
 addDefaultEmailLogger <- function(mailSettings, label = Sys.info()["nodename"], test = FALSE) {
-  registerLogger(createLogger(name = "DEFAULT",
+  registerLogger(createLogger(name = "DEFAULT_EMAIL_LOGGER",
                               threshold = "FATAL",
                               appenders = list(createEmailAppender(layout = layoutEmail,
                                                                    mailSettings = mailSettings,
                                                                    label = label,
                                                                    test = test))))
+}
+
+#' Add the default error report logger
+#'
+#' @details
+#' Creates a logger that writes to a file using the "FATAL" threshold and the
+#' \code{\link{layoutErrorReport}} layout. The file will be overwritten if it is 
+#' older than 60 seconds. The user will be notified that the error report has been
+#' created, and where to find it.
+#'
+#' @param fileName   The name of the file to write to.
+#'
+#' @export
+addDefaultErrorReportLogger <- function(fileName = file.path(getwd(), "errorReportR.txt")) {
+  registerLogger(createLogger(name = "DEFAULT_ERRORREPORT_LOGGER",
+                              threshold = "FATAL",
+                              appenders = list(createFileAppender(layout = layoutErrorReport,
+                                                                  fileName = fileName,
+                                                                  overwrite = TRUE,
+                                                                  expirationTime = 60)))) 
 }