+++ R/ParallelLogger.R
@@ -20,5 +20,7 @@
 "_PACKAGE"
 
 #' @importFrom methods getFunction is
-#' @importFrom utils installed.packages install.packages menu capture.output help memory.limit setTxtProgressBar txtProgressBar
+#' @importFrom utils installed.packages install.packages menu capture.output 
+#'             help memory.limit setTxtProgressBar txtProgressBar limitedLabels 
+#'             sessionInfo
 NULL